build es-de instead of installing binary.
Building it is better, I guess, for source provenance (whatever it is)(https://ryantm.github.io/nixpkgs/stdenv/meta/#sec-meta-sourceProvenance). It also allows to remove installPhase (it will be done automatically) and set some compile flags (remove application updater, for example). Added myself as maintainer because I'm interested in this package (feel free to remove me if that's a problem).